Add external linked css style sheet
New style.css file added to the main directory level
and linked to in head of index and content html files.
This styling is simple practice and uses type selectors
for speed, rather than going back over each html element
and adding consistent named classes.
One exception is an all-encompassing div with the class
name of "main" which was an easy add directly nested under
the body tags.